+++ src/test/kotlin/org/jetbrains/intellij/IntelliJPluginSpecBase.kt
@@ -62,7 +62,7 @@ abstract class IntelliJPluginSpecBase : IntelliJPlatformTestBase() {
             """
             plugins {
                 id 'java'
-                id 'org.jetbrains.intellij'
+                id 'org.jetbrains.intellij.platform'
                 id 'org.jetbrains.kotlin.jvm' version '$kotlinPluginVersion'
             }
             sourceCompatibility = 11